优化时间戳

+++ sources/prediction.move
@@ -17,8 +17,8 @@ module suiprediction::prediction {
 
     struct Round has  store {
         epoch: u32,
-        lockTimestamp: u32,
-        closeTimestamp: u32,
+        lockTimestamp: u64,
+        closeTimestamp: u64,
         lockPrice: u128,
         closePrice: u128,
         totalAmount: Balance<SUI>,
@@ -99,7 +99,7 @@ module suiprediction::prediction {
     public entry fun startplay(rounds: &mut Rounds,epoch: &mut Epoch,clock: &Clock,ctx: &mut TxContext) {
         let firstround = Round {
             epoch: 0,
-            lockTimestamp: (clock::timestamp_ms(clock) as u32),
+            lockTimestamp: clock::timestamp_ms(clock),
             closeTimestamp: 3333,
             lockPrice: 0,
             closePrice: 1,
@@ -133,7 +133,7 @@ module suiprediction::prediction {
         assert!(value > 0,0);
 
         round.closePrice = value;
-        round.closeTimestamp = (clock::timestamp_ms(clock) as u32);
+        round.closeTimestamp = clock::timestamp_ms(clock);
 
         // up wins
         if(round.closePrice > round.lockPrice) {
@@ -150,7 +150,7 @@ module suiprediction::prediction {
         };
         let nextround = Round {
             epoch: (epoch.currentEpoch + 1 as u32),
-            lockTimestamp: (clock::timestamp_ms(clock) as u32),
+            lockTimestamp: clock::timestamp_ms(clock),
             closeTimestamp: 0,
             lockPrice: value,
             closePrice: 0,
@@ -392,6 +392,6 @@ module suiprediction::prediction {
 }
 
 
-//   - ID: 0x21d68b314dc7ac0204fca16c2c0dd19538b2304976d4dec6880ee8fc65d5b8ce ,  rounds
-//   - ID: 0x9b82c157fc3959d75c8a6beed9caf19c8fae3bffa1ad2b8d858315cb4556697c ,  Immutable
-//   - ID: 0xbe3200d74f26392eb5ae78e7ba9eec3a2b5c4f50ae7b20881f873ded140f36fb ,  epoch
+//   - ID: 0xe88e990f1735dfdc241dbb317d871922daa9f78e42d264cbe1ec9aa4793ccd7c ,  rounds
+//   - ID: 0xfd4c3da227b886abba9ddbf82eef2d72f248f0f1c8b572d3d735dfc51fb1ece8 ,  Immutable
+//   - ID: 0xad3b80dae522ed6cc545f8548b72bc5569551856c4d93fbd222a0c9f20acdf10 ,  epoch
